Calling it a night
I have the form set up. I also have the ability to both push
information to my database as well as pull information from my
database. I need to tweak this form so that it can be used for what I
need it for. I now have one page created and now just need to finish
the other three.