Blog Changes
I made some changes to the blog.
There are now pages for both the Admin view (blog create)
and the user view (add comments)  I just have to fix the chat, the user
profile and the home page and I’m done.   And the chat….It’s still not
connecting to my database for some reason.